# src/api/Swagger/index.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Recipe Book API",
    "description": "Recipe Book API documentation using swagger.",
    "contact": {
      "email": "mattia.buffetti@gmail.com"
    }
  },
  "servers": [
    {
      "url": "https://recipebook-lemon.vercel.app",
      "description": "Production server"
    },
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    }
  ],
  "basePath": "/api",
  "paths": {
    "/auth/register": {
      "post": {
        "tags": [
          "User"
        ],
        "summary": "Create a new user",
        "description": "Create a new user",
        "produces": [
          "application/json"
        ],
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "user",
            "description": "The user to create.",
            "schema": {
              "type": "object",
              "required": [
                "email"
              ],
              "properties": {
                "name": {
                  "type": "string"
                },
                "email": {
                  "type": "string"
                },
                "password": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User saved successfully",
            "schema": {
              "type": "object",
              "$ref": "#/definitions/User"
            }
          },
          "400": {
            "description": "Error in saving user",
            "schema": {
              "type": "object",
              "$ref": "#/definitions/CommonErrorResponse"
            }
          },
          "500": {
            "description": "Error in saving user",
            "schema": {
              "type": "object",
              "$ref": "#/definitions/CommonErrorResponse"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "CommonErrorResponse": {
      "type": "object",
      "properties": {
        "error": {
          "type": "object",
          "properties": {
            "code": {
              "type": "number"
            },
            "message": {
              "type": "string"
            },
            "keyValues": {
              "type": "object"
            }
          }
        },
        "success": {
          "type": "boolean"
        }
      }
    },
    "User": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "role": {
          "type": "string"
        }
      }
    }
  }
}